changing directory again :(
