+++ The-Fork-New-Website/package.json
@@ -20,6 +20,7 @@
     "react-bootstrap": "^1.0.0-beta.14",
     "react-dom": "^16.10.2",
     "react-icons": "^3.7.0",
+    "reactstrap": "^8.1.1",
     "scrollreveal": "^4.0.5"
   },
   "devDependencies": {
