+++ The-Fork-New-Website/src/pages/index.js
@@ -8,11 +8,13 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 // Page Components
 import Footer from '../components/Footer/Footer'
 import About from '../components/About/About'
+// import Team from '../components/Team/Team'
 
 // Style
 import './style.css'
 
 export default () => <div>
+    <Navbar />
     <About />
     <Footer />
 </div>
