+++ The-Fork-New-Website/src/pages/style.css
@@ -1,3 +1,7 @@
+:root {
+    --primary:#026666;
+}
+
 * {
     padding: 0;
     margin: 0;
