--- netlify.toml
@@ -1,31 +1,0 @@
-# Settings in the [build] context are global and are applied to all contexts
-# unless otherwise overridden by more specific contexts.
-[build]
-  # Directory to change to before starting a build.
-  # This is where we will look for package.json/.nvmrc/etc.
-  base = "/The-Fork-New-Website"
-
-  # Directory (relative to root of your repo) that contains the deploy-ready
-  # HTML files and assets generated by the build. If a base directory has
-  # been specified, include it in the publish directory path.
-  publish = "public/"
-
-  # Default build command.
-  command = "gatsby build"
-
-# Allows CORS
-[[headers]]
-  for = "/"
-  [headers.values]
-    Access-Control-Allow-Origin = ""
-
-# Production context: all deploys from the Production branch set in your site's
-# deploy contexts will inherit these settings.
-[context.production]
-  publish = "public/"
-  command = "gatsby build"
-
-# Deploy Preview context: all deploys generated from a pull/merge request will
-# inherit these settings.
-[context.deploy-preview]
-  publish = "public/"
